+++ plugins/AppTrade/hhvm/AppTrade.php
@@ -24,7 +24,7 @@ abstract class AppTrade {
 	(
 		int $senderID		// <int> The UniID sending the Auro.
 	,	int $recipientID	// <int> The UniID receiving the Auro.
-	,	float $auroAmount		// <float> The amount of Auro being sent.
+	,	int $auroAmount		// <int> The amount of Auro being sent.
 	,	string $desc = ""		// <str> The description for the log.
 	,	bool $anon = false	// <bool> Whether to show it as anonymous or not.
 	): bool					// RETURNS <bool> TRUE if the Auro was sent, FALSE if it failed.
